Add Tutorial System
A widget sitting in the cockpit gives tutorials for driving, flying, and landing the airplane. DIfficult words are defined as well. The tutorial is called during key events: game startup, flight, and when the gps finishes. Tutorial is completely optional and can be disabled when the game is started
As a result of this implementation the GPS is also completed. When the start point and end point contact each other, the start point is removed and the end point is turned green to signify your destination that you've reached and can see. 
I also attempted to fix the z-fighting issue causing the GPS map and the player's map to collide and cause issues